Only can view and remove black list entried when you're part of the team

+++ app/controllers/black_lists_controller.rb
@@ -1,10 +1,14 @@
 class BlackListsController < ApplicationController
+  after_action :verify_authorized, except: :index
+  after_action :verify_policy_scoped, only: :index
+
   def index
-    @black_lists = BlackList.all.order(created_at: :desc).page(params[:page])
+    @black_lists = policy_scope(BlackList).order(created_at: :desc).page(params[:page])
   end
 
   def destroy
     black_list = BlackList.find(params[:id])
+    authorize black_list
     black_list.destroy
     redirect_to black_lists_path
   end
